requirements.txt
autoenv==1.0.0
Flask==1.1.1
gunicorn==19.9.0
markdown2==2.4.0
Pillow==6.1.0
python-dotenv==0.19.0
requests==2.22.0
